# Generate ids with itertools.count