Avoids the critical section with threading.Lock in favor of itertools.count.

`count` objects are threadsafe, and their critical section is implemented in C and provide better performance that Python level locking.

### Preliminary Benchmarks

This update also provides some speedup, I will make no claims on the factor as the benchmarks are not rigorous. 
Setup: Just ran these on my M2 macos with Python 3.12.5, better benching methodology can be done if requested.

**After**

```bash
$ python -m timeit -s "from pipecat.utils.utils import obj_count, obj_id" "obj_id()"
10000000 loops, best of 5: 31.7 nsec per loop
$ python -m timeit -s "from pipecat.utils.utils import obj_count, obj_id" "obj_count(object)"
5000000 loops, best of 5: 93.8 nsec per loop
```

**Before**

```bash
$ python -m timeit -s "from pipecat.utils.utils import obj_count, obj_id" "obj_id()"
2000000 loops, best of 5: 130 nsec per loop
$ python -m timeit -s "from pipecat.utils.utils import obj_count, obj_id" "obj_count(object)"
1000000 loops, best of 5: 226 nsec per loop
```